Menu Refactor
Unnötige MenuBarPanel Klasse entfernt. LoginListener zum KontoMenu hinzugefügt um die Sichtbarkeit bei den MenuItems zu kontrollieren
